Use Hash based routing for project
- Since this is hosted on GitHub Pages and therefore doesn't have all
of its routes being served by the Ember app running on the server, means
that when we try to navigate to any route that isn't the default, GitHub
Pages will try and find an html file in the Pages branch with the route
name. That'll never exist, since this is a single-page application. To
get around that issue, we switch to using Hash based routing. That
changes the URLs for subroutes from looking like
https://mydomain.com/subroute to https://mydomain.com/#/subroute.
Note that new "#" in the URL. Since the hash part of a URL isn't
something that's used for navigation by the browser (i.e., browsers
don't consider hash changes to be something that should attempt to hit
a server to fetch a new file), the Ember app stays loaded and will read the
hash change in order to navigate to a new child route.

This specifically fixes an issue where the "your-card" route would fail
to load in a new window, but will also keep that issue solved for any
future child routes.

See https://api.emberjs.com/ember/release/classes/HashLocation for more
information.